+++ packages/styleguide/src/styles/breakpoints.ts
@@ -1,0 +1,5 @@
+export const breakpoints = {
+  small: 400,
+  medium: 900,
+  large: 1200,
+};