+++ packages/styleguide-native/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.2-alpha.1](https://github.com/expo/styleguide/compare/@expo/styleguide-native@0.4.2-alpha.0...@expo/styleguide-native@0.4.2-alpha.1) (2022-01-05)
+
+**Note:** Version bump only for package @expo/styleguide-native
+
+
+
+
+
+## [0.4.2-alpha.0](https://github.com/expo/styleguide/compare/@expo/styleguide-native@0.4.1...@expo/styleguide-native@0.4.2-alpha.0) (2021-12-26)
+
+**Note:** Version bump only for package @expo/styleguide-native
+
+
+
+
+
 ## [0.4.1](https://github.com/expo/styleguide/compare/@expo/styleguide-native@0.4.0...@expo/styleguide-native@0.4.1) (2021-12-09)